Add Float16, Float32 and Float64 (AbstractFloat) support.

+++ test/test_model.jl
@@ -7,8 +7,8 @@ using LatticeBoltzmann
 	@test length(model.domains) == 1
 	@test length(model.domains[1].ρ) == 256^3
 
-	@test model.ρ[1] == 1.0f0
-	@test model.u[1, 1] == 0.0f0
+	@test model.ρ[1] == Float32(1.0)
+	@test model.u[1, 1] == Float32(0.0)
 	@test model.flags[1] == 0x00
 
     @test model.domains[1].ρ === model.ρ.buffers[1]
